Исправить обновление количества комментариев в карточке
